refactor: init module
